Cleanup products package structure and file names
