+++ feature/products/src/main/java/co/ke/xently/products/ProductsActivity.kt
@@ -44,10 +44,10 @@ internal fun ProductsNavHost(
         composable("products") {
             ProductListScreen(
                 modifier = Modifier.fillMaxSize(),
-                onNavigationIconClicked = onNavigationIconClicked,
                 onItemClicked = {
                     navController.navigate("products/$it")
                 },
+                onNavigationIconClicked = onNavigationIconClicked,
             )
         }
     }
